Create Day 5
Explaination

This Python script asks the user to input two values, which represent the number of boys and girls, separated by a space. It uses the input() function to take the user's input, and the split() method is then used to separate the input string into two distinct values. These values are assigned to the variables x and y, representing the number of boys and girls, respectively. The script then prints out the number of boys and girls by outputting the values of x and y. This code demonstrates basic user input handling, string splitting, and variable assignment in Python.