Create Day 6
Explaination

This Python script demonstrates how to check the data types of different variables. The variables a, b, c, d, e, and f are assigned various types of values:

a is a string ("Hello chapri")

b is an integer (-10)

c is a floating-point number (3.14)

d is a tuple (("geeks", "for", "geeks"))

e is a list (['a', 'b', 1, 2, 3])

f is a dictionary ({1: 'a', 2: 'b'})

The type() function is then used to print the data type of each variable, which outputs the type of the object each variable is holding. This demonstrates how to check the data type of different Python objects.