Create Day 5
code for Bandname generator 
explaination
This simple Python script is a fun "Band Name Generator" that asks the user to input the name of the city they grew up in and the name of a pet. It begins by printing a welcome message using the print() function. Then, it uses the input() function to collect the city name and pet name from the user. After both inputs are received, it combines them with a space in between and displays the result as a possible band name using another print() statement. The goal is to create a creative and personalized band name based on user input.