Create Dayy 6
Explaination
code1: This code prints numbers from 0 to 10 using a for loop. The variable n is set to 11, and range(0, n) generates a sequence of numbers starting from 0 up to (but not including) 11. In each iteration of the loop, the current value of i is printed. As a result, the output will be a list of integers: 0, 1, 2, ..., up to 10.

code2: This code iterates through each character in the string "sankettt" and prints them one by one. The loop uses the variable char to temporarily hold each character from the string during each iteration. This is a common way to process or analyze individual characters in a string in Python. The output will be each letter of the word sankettt printed on a new line.